Move services folder under src
