+++ src/events/guildCreate/index.js
@@ -1,6 +1,6 @@
 const { WebhookClient } = require('discord.js');
 const { webhooks } = require('../../../config/yagi.json');
-const { insertNewGuild } = require('../../../services/database');
+const { insertNewGuild } = require('../../services/database');
 const { sendErrorLog, serverEmbed, checkIfInDevelopment } = require('../../utils/helpers');
 
 /**
