+++ src/services/database.js
@@ -1,5 +1,5 @@
 const { Pool } = require('pg');
-const { databaseConfig } = require('../config/database');
+const { databaseConfig } = require('../../config/database');
 const pool = new Pool(databaseConfig);
 
 exports.createGuildTable = async (guildsOfYagi, yagi) => {
